requirements.txt
requests>=2.28.0
plyer>=2.1.0
matplotlib>=3.5.0
pystray>=0.19.0
pillow>=9.0.0
numpy>=1.22.0
pyinstaller>=5.6.0
